Выполняет задание по модуляции
